agents/party/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-party"
version = "0.1.0"
description = "Party voice agent — chained STT/LLM/TTS baseline for multi-agent room"
requires-python = ">=3.11,<3.13"
dependencies = [
    "sdk",
    "tools",
    "livekit-agents[openai,google,silero]~=1.5",
    "livekit-plugins-elevenlabs~=1.5",
]

[tool.uv.sources]
sdk = { workspace = true }
tools = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=1",
]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.setuptools.packages.find]
where = ["src"]
